[REFACTOR]: Continuing the refactoring of code to TS
In this commit, the following changes have been made:

*** ADDED ***
- `interfaces/user.ts`: Interfaces for logged-in user, anonymous user, and conversational partner have been added. These interfaces are designed for use in `UserStore.ts` and `StrangerProfileStore.ts`.
- `types/user.ts`: Types for actions performed on a user in `UserStore.ts` have been added.

*** MODIFIED ***
All components and Pinia state management have been modified to utilize TypeScript. However, not all of them function correctly at this stage.
This is intentional, and these issues will be addressed in the immediate next commit.

In `UserStore.ts`, the `performUserAction` function has been expanded to accept action type and data for executing the action.
Subsequently, based on the action type, the appropriate function from `UserStore.ts` is invoked using a switch case.
This change has been introduced to simplify the management of actions performed on the user and reduce code redundancy.

Modules currently functioning correctly:
- All from the previous commit,
- All components from the `modals` folder,
- `StrangerProfile.vue`,
- `StrangerTitle.vue`,
- `DesktopNavbar.vue`,
- `MainNavbar.vue`,
- `SettingsNavbar.vue`,
- `RangeSlider.vue`,
- `UserStore.ts`,
- `setAvatar.ts`

NOTE:
This commit represents a step forward in transitioning the project to TypeScript and improving its maintainability.